requirements.lock.txt
# Compiled lock (pin versions for reproducibility)
fastapi==0.110.0
uvicorn==0.27.0
pydantic==1.10.13
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
scipy==1.12.0
onnxruntime==1.23.1
orjson==3.9.15
httpx==0.27.0
redis==5.0.1
motor==3.5.1
prometheus-client==0.20.0
pyjwt==2.8.0
tenacity==8.2.3
psutil==5.9.8
